pyproject.toml
[tool.poetry]
name = "scenario-transferring-visualization"
version = "0.1.0"
description = ""
authors = ["changnam hong <hcn1519@gmail.com>"]
readme = "README.md"
packages = [{include = "scenario_transferring_visualization"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.2"
plantweb = "^1.2.1"
pyyaml = "^6.0.1"
pytest = "^8.1.1"
cyber-record = "^0.1.12"
protobuf3-to-dict = "^0.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
